refactor: Split AccountTab into smaller components
